`tools/importer-rest-api-specs`: moving the Terraform Test Generation logic into the Terraform package
