+++ tools/importer-rest-api-specs/components/terraform/testing/dependencies_template.go
