+++ tools/importer-rest-api-specs/components/terraform/testing/variables_template.go
