+++ tools/importer-rest-api-specs/components/terraform/testing/variables_template_test.go
