+++ tools/importer-rest-api-specs/pipeline/task_generate_terraform_tests.go
@@ -5,9 +5,9 @@ package pipeline
 
 import (
 	"fmt"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/terraform/testing"
 
 	"github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/testing"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/models"
 )
